"Correct variable name and change variable type for VPN check"
Updated the variable name from 'chec_vpn' to 'check_vpn' in 'NeoxGeoBag.php', 'neoxBag.php', 'NeoxGeoBagService.php', and 'Configuration.php' files to fix a typographical mistake. Also, the type of 'check_vpn' was updated from 'bool' to 'string' in 'neoxBag.php' file to match correctly with the 'setCheckVpn' and 'getCheckVpn' function signatures. The changes were done for better code readability and consistency, as well as to correct the type confusion error that can occur during function calls related to vpn checks.

In addition, a code block of timer setting that was commented in 'geolocatorAbstract.php' doesn't affect the program, as geolocation is still functioning as expected. The code block can potentially be removed in the future, further simplifying the code.