"Change data type of 'check_vpn' variable from bool to string"
The 'check_vpn' variable used in the configuration of NeoxGeoBag was previously a boolean. However, it was causing type confusion errors during function calls related to VPN checks, since the 'setCheckVpn' and 'getCheckVpn' functions expected a string. To resolve this discrepancy, the data type of 'check_vpn' has been changed from bool to string. This rectifies the error and improves code consistency and readability.