Add ping IP check in GeoLocatorSubscriber and implement rate limiter
Enabled 'onKernelRequest' in 'GeoLocatorSubscriber' to check if the request is in banned state by using 'checkIpPing' from 'geolocatorAbstract'. It throws 'BadRequestHttpException' in case of banned IPs. Added 'getIpPing' in 'Limiter' to implement request rate limiting and clear cache if the limit is exceeded where the limit is set as 5 and the cache expiry is 10s, after the limit is exceeded the limit is reset to 20 and the cache expiry is extended to 600s. The change was necessary to mitigate any security risk because of abusive or suspicious IP requests.