Updated neoxBag for ping checks and improved geolocator validation
Introduced a new 'check_ping' attribute in 'neoxBag' to enable/disable ping checks. This is set in 'NeoxGeoBag' and used in 'GeoLocatorSubscriber', which throws an exception for banned IPs. Improved request rate limiting in 'Limitor' by considering this new attribute. If requests exceeded the limit of 5 within 10s, the counter cache is cleared and the request limit is increased to 20 for 600s. This change enhances security by preventing abusive access from suspicious IPs.