Refactor functions in multiple services
Updates to three files are included in this commit. In `Configuration.php`, a function argument has been corrected. In `NeoxGeoBagService.php`, a helper function `toCamelCase` has been added to ensure proper method-naming convention when setting values. In `GeoLocatorSubscriber.php`, the `containsKeyword` function has been refined to not only check for the presence of a keyword but also ascertain its position in the string.